refactor: lift tracing to the peer connection wrappers
